requirements.txt
opencv-python
torch
torchvision
tensorflow
mtcnn
matplotlib
tqdm
kaggle
mxnet
scikit-learn
pillow
